Added AuthGuard for the "update"- and "destroy"-routes

+++ app/Http/Controllers/PostsController.php
@@ -132,6 +132,12 @@ class PostsController extends Controller
 
         //  Update Post
         $post = Post::find($id);
+
+        //  Check for user authentication
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+        
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -148,6 +154,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //  Check for user authentication
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
